Finished model branch, commit, commitID, also for-each-ref command
This should be a skelton that is enough for me to work on branch next.
Added a few changes to the files listed bellow.
I have to find the reason behind the weird duplicate model in branch.
I also need to write a test for for-each-ref.
Any length checking for the names of branches or repos is left to the
caller.
# On branch master
# Your branch is ahead of 'origin/master' by 1 commit.
#   (use "git push" to publish your local commits)
#
# Changes to be committed:
#	deleted:    build/main.js
#	deleted:    build/main.js.map
#	modified:   index.html
#	new file:   src/git-drive/for-each-ref.ts
#	modified:   src/git-drive/index.ts
#	modified:   src/main.ts
#	modified:   src/model/branch.ts
#	modified:   src/model/commit.ts
#	modified:   src/model/repository.ts
#	new file:   src/util/remote-prefix.ts
#	modified:   tsconfig.json
#