added the first test, yay!
The tests will be made using Mocha and Chai (very jittery I know).
The first test is for the command for-each-refs.
I need to test large amounts of branches.
I also need to make the scripts that create repos and delete them with
every test. More robust in the sense that I don't need to hardcode a
path into the tests.

# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   package-lock.json
#	modified:   package.json
#	new file:   tests/git-drive/for-each-ref.ts
#	new file:   tests/testRepos/twoBranches
#
# Changes not staged for commit:
#	modified:   build/main.d.ts
#	modified:   src/git-drive/for-each-ref.ts
#