Very basic implemntation of push/pull, baiscally figured out the options
I also realized another potential problem. People can commit using
GitHub, very bad and I wrote a shpeal about it in pull. We will have a
bad time trying to deal with that or most likely not be able to deal
with it. Baiscally if anyone commits on GitHub and breaks our rules we
are in trouble. Progress should come really soon, I just want to finish
a skeleton of most of the model by end of next week.

# On branch master
# Your branch is up to date with 'origin/master'.
#
# Changes to be committed:
#	new file:   src/git-drive/pull.ts
#	new file:   src/git-drive/push.ts
#
# Untracked files:
#	src/git-drive/fetch.ts
#