Added the diff-index command as it will get useful
The command is useful in knowing if something is staged or not. As it
will show the file when it is staged and won't if it is not. Something
that we do not have no explicit support for as of now. I do not think we
will need it but its better to have a way than not.