Re-arranged files and finally added path mapping to the modules
Cool stuff, I need to figure out how to export the git function in
git-core as a module without declaring a class for it (I know it sounds
silly) but I still need to reference the file in a relative way because
of that. For now it is the only exception to the rule.