Changed the imports for the tests
So I guess fuck mocha. It seems that I can't easily point mocha to my
tsconfig file for it to understand the path mappings. Which means that
the tests will have to have ugly imports. There is a hack around it but
I don't feel like it is worth spending the time to understand and setup
for now. Read this for more info: https://github.com/TypeStrong/ts-node/issues/319