Added the POST sytem interfaces and enums
This is a nice way to interact with the post system. So if a the variant
passed is going to be either or we can just use the type variant then
check for the type, this type checker is aware of this fully too so it
will warn us if we make assumptions about the type. Or if we are sure we
can just use the interface directly. main.ts has an example of how to
use this at the end of the on.(ready... function.