So we begin, adding our classes and methods
So far so good, there is a lot to be done and a lot more to figure out.
I have to find a way to add mentioning what files a commit changes into
the log, doing this command commit by commit is not gonna be a good way
to go about it. Also partialcheckouts is WIP atm. This has path mappings
as well for the new files I added.