I finally have full on path mapping and aliasing in compile and runtime
It turns out typescrip doesn't write the resolved path mappings in the
compiled code, it only uses it to check at compile time. I had to get a
temp dependency that overrides the behavoir of require to include the
mappings that I had to define again in the package.json with reference
to build (as opposed to src in tsconfig). This looks good for now till I
get a build tool that will resolve tha mappings.