Made a decision to include undefined into the return type of getVal
So this will enforce the called to check whether the value they passed
for the key exists without them needing to use try and catch. If this
wasn't the case the caller has might forget that the function could
error out as the type checks out until it does in runtime. I do not
think I'll need the function without the undefined as a return type, if
it I do, I will just make the current one be warpper to the original.