General edits here and there
Cleaned things up in user, added IChangeInterface for user too.
Explained why parialcheckout is not as strict, since it is harmless
without being able to commit and our committing is very strict.
started start lol. So much will go into it that I will do it later as we
have to have an agreement on what to do.