So I implemented statRepo method
It took a bit of work to figure this out but so for so good. I need to
comment it and clean up the code a bit here and there. I also would want
to pass the function a current user (or use a global var) to checkout
thier branch at the end of the method.