Changed models to the better and finished up more of start and sync
So now start checks if a repo already exists before and throws an error
if it does. It also creates the correct user objects for the topicspace.
I removed the user object from workspace as it is redundant. I added the
first commit obj to topicspace as it will be useful in creating new
workspaces.