Added some error handeling
So far this is a rough error handeling scheme, it mainly relies on the
errors that dugite understands how to parse. I also removed useless code
from renamBranch and added my check to rev-parse. I need to find out why
dugite does not throw the error in rev parse as it should.