Just a quick fix that the compiler complains about now for some reason
This is actually intersting, I will look more into this later, once the
function call that returns never went away the compiler realized that
the code under the if statement is dead. I knew it was before but I just
kept it since vs code complained about how I have to return a value if I
took it out.