So I added the the add topicspaces command along with some other updates
I hard coded the fact that the branch that has to be moved by
createworkspaces function has to be called master, I changed that to let
the caller decide on what banch they want to move. I also made the id
for G-repo look a little bit better. Finally update the addTS to include
users in at the repo level that were not part of the repo before.