commented and cleaned, I also found a problem with creating TSs
So the create TS functionality records where the new TS is based but the
problem with that is that the caller is passing us that commit and we
are not doing anything with it. In other words, the caller could have
another branch checkedout and is passing us some other commit. Since we
create the new TS based on HEAD's state this could lead to us keeping
false info about the new TS. The solution is to provide the orphan
checkout with the SHA of the commit the caller passes to us. Will do
this at some other time as I don't have time now.