Added the first stage to loading a repo form disk
This is the first step of verifying the structure of the repo to later
match it with the .proto file. Since these tests pass, we know the repo
objects are created following our structure at least. We will not verfiy
commits on branches except for the tip and the first. This is left for
verfiy method as it is resource consuming.