Switched to protubuf js by google, added proto files with their code
So we decided to make the switch to protobuf made by google as it makes
more sense to do so, plus it is closer to actual protobuf and is more
popular than protobuf.js. All these files should be generated before
hand and shipped with the code, thus why they are being committed. Any
changes in them should be reflected in their models.