These are our objects wrapping around protobuf
These are our models warpping around protobuf while extending/building
on the basic one at the same time. It is nice since they dep graph goes
one way, meaing basic objects can be created and saved in proto if we
need to. I need to comment and test this more, "it seems" like it is
working for now. I also need to clean up a bit. I will have to explain
what I did in the code as well. Also these are not the final state of
the models, more changes to come. I will add a map object to workspaces
once the GFile is implemented better, for now this works.