Made some changes to the commands/scripts based on new proto models
The changes have to do with the fact that the user object keeps an array
of workspaces now instead of a key value pair. Sadly key value pairs are
not useable in protobuf, so we are forced to use the map class (not
really a problem, it was just wasted work on my side, meh).