This is a working example of a serialze and deserialze of proto obj
As the title suggests. I also fixed a really interesting bug the
happened after the proto integration. I was passing old pointers or
workspaces and topicspaces into addworkspace. Before this was not a
problem (I don't really know why) but now the actuall state of the
object doesn't update correctly, only the new change gets updated, the
old one gets overrwritten since the function uses the old one that is
passed. This is really intersting, so I either have to find a way to
reference TSs and WSs in the repo object without actually passing
references to them. This will increase the ease of use and minimize
error as the repo object will be the sole keeper of state. Or I would
have to match the names (or ids, if we use any) and from there find the
TSs and WSs through the repo object and use those instead of the ones
passed. I don't like the second solution but we'll see, these are my
inital thougths, more thinking tomorrow.