This should conclude the protobuf integration for now
There are things left to do, more testing, but for now this works. I
commented and cleaned up the code, plus looked over it again to fix
anything missing. One thing might be is to change the top level arrays
in GRepo into a map<string, Topicspace> and map<string, User>. This
should not change much of the surface code but it will make interacting
with the repo objects easier as the user will not be passing
pointers/references around instead they will pass unique string
identifiers.