So I figured out the problem I was having friday
After a while of testing and losing my mind, I managed to figure out the
problem had nothing to do with protobuf or how I wrapped around it. The
problem was with the read and write node calls. Apparently the code will
go on and not wait for the write call to end even though I used await
and the code is async. This is stupid I know, so I now use fs-extra
which promisfies all the fs calls so I am granteed that my code will
fucking sit there and wait for the write to fully happen. This should
actually and hopefully end my protobuf struggle so I can move on to
actually impelementing the rest of the app commands.