This loading of repos is coming along slowly
So the metadata is proving to be very hard to do nicely and simply.
There is the problem that we have to do the book keeping before
committing, meaning that the info the commit is introducing is not gonna
be part of what we are tracking. This gets really problematic when there
are new workspace or topicspaces added to the repo as we cannot have
that info recorded in other branches. This means that unless we are
reading the correct workspace (branch) we might not know that these
workspaces or topicspaces exist. I mean we know from the structure of
the repo, I already have code that does that, but we don't know the
metadata and we have to find or look for all of it again ourselves. I
want to avoid having a global branch that keeps track of the state of
the repo globaly as doing so will need a checkout very often. Maybe I
will do one that only requires a checkout when creating a ts or a
workspace, we'll see on this idea.