Added the change in the Grepo to where it need to be added
This is not bad so far, we only need to keep the name the meta branch,
not the tip or anything else. I also realized that if we are creating
workspaces we do not need to have a GRepo obj, only a repo which easier
to create. I have to integrate this in the other structure changes
commands.