Finished up working on load
God this took forever but its worth it. It is not fully tested but it
works as intended. I think this is decent for now. I will find other
stuff to add and will have to figure out more stuff once everything is
implemented. For now we can move on.