Started writing test for our app commands
Always a fun process. I also had to remove the tsconfig-paths from the
mocha command. Writing these made me think and a find a few bugs so far.
More to come monday and I have to take the error checking approach to
the lower git commands since dugite might not have an error that we care
about (I doubt it, but we will see).