Fianlly wrote tests for the app commands
So here we have two tests for each command, one to check that it is
working as expected and one to check that it error as expected. I like
this way of testing and once error handeling for git command is fully
functional I should test them out like this as well. There are a few
error tests for load left to do. They are the test that will require me
to create invalid repos, later will do for them.