Just added the basic grid for the app
The app will have three areas, the first is the menue on the top, it
will contain certain commands like sync, load, download, and start.
The second is the topicspaces and workspaces pane, here the user should
be able to navigate between the the topicspace and workspaces in what
looks like a file system tree. Lastly it is the file explorer in which
the user should be able to right click a file to view its history and
open stuff, etc...