Clean up the dependencies folder and added the vuex along with vuex-ts
vuex-typescript will be used to create the store without the headache of
having to define it in the verbose way normal vuex asks us to do. Sadly
the vue-class-component has to be there and has to be used. This is fine
for now but we will see if it changes later.