Added the store files and changed App.vue to be class component based
So this is how things will probably play out, I really like the vuex
store idea of mutations. This does everything correctly for now except
that the computed property fTime in App.vue is undefined for some
reason. I might need to move dispatcher up a level for this to work, I
will see. I also have to move the interfaces into their own file
(maybe). So far so good.