Added the rest of the componenets along with disabling strictPropertyInitialization
So yeah these will be our main three components. I only use one of them
now but the rest should follow the same way. The choice of making TSs a
prop in the twssPane is not final, also the names are not final as well.
I disable the option strictPropertyInitialization in the compiler
because typescript would complain otherwise about props not being
initalized. I also should have had the index.ts be included in two
commits down.