Added a couple of state changes and fixed how state is mutated.
Beforehand we could not mutate the state since everything was a
readonly. The only thing we need to make sure mutates and is consistant
is what we will be writting to desk, ie. appData. So that got fixed and
there will probably be more things in appState that are private but not
readonly, making them changable within the same instance of appState.
appData is the only object we recreate in other words. The divide
between the two might seem strange but dispatch might and will probably
have to deal with more state things than appState and update them
accordingly. We will see.