Update the components overall and link some funcitonality
So this is still barebone but we have the add repo function sorta
working. The UI is very minimal, I need to find a color scheme and stick
to it, better than the green stuff. I also will figure out how to do the
create repo functionality, I think I will need another browserwindow for
that no matter what. Things are still and will move around a lot.