Did some clean up and added some listeners
So I dug into the whole idea of creating popups as dialogs or modals but
it all seems like a can of worms we do not want to deal with. I also
looked into adding bootstrap and bootstrap-vue to the project it did not
work and I have no clue why. Fuck it we don't need bootstrap as the cool
stuff needs jquery and no we are not adding jquery to the tech stack.