So I figured out that tests do not run since they need a window
After a long fight I decided it is not worth it to do the tests for
oauth stuff. I can but it will be a struggle. To compensate I have a
nice example in examples. It cover all the functions we need to use from
the github api. Nothing too complicated. One thing worth mentioning is
some of the objects returned are reduced, meaning that not all of the
object members are exposed in typescript interface definitions. This is
not a problem but something worth looking at just to know what kind of
info is there.