Fixed a bug with repo names
So when we write the repo proto files we write the path and name of the
repo. This is wrong because at least one of them will change upon
someone else cloning and reading the repo. So what we do now is save the
path as "-/-" and name will be as a result "-". We can abstract this away
since we know the path we will write to or read the proto file from.
This implies that the basename of a repo will always be the name of the
repo, which is the case indeed.