This is the most up to date version of the documentation
I will finish this up at some point in the next week or so. I just can't
do it under such a time rushThis is the most up to date version of the
documentation

I will finish this up at some point in the next week or so. I just can't
do it under such a time rush